+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/command/WandCommandExecutor.java
@@ -148,6 +148,7 @@ public class WandCommandExecutor extends MagicConfigurableExecutor {
 			addIfPermissible(sender, options, "Magic.commands." + permissionKey + ".", "unbind");
 			addIfPermissible(sender, options, "Magic.commands." + permissionKey + ".", "save");
 			addIfPermissible(sender, options, "Magic.commands." + permissionKey + ".", "delete");
+			addIfPermissible(sender, options, "Magic.commands." + permissionKey + ".", "levelspells");
 
 			Collection<String> allWands = api.getWandKeys();
 			for (String wandKey : allWands) {
@@ -464,6 +465,13 @@ public class WandCommandExecutor extends MagicConfigurableExecutor {
 			if (!api.hasPermission(sender, "Magic.commands." + command + "." + subCommand)) return true;
 
 			onWandOrganize(sender, player);
+			return true;
+		}
+		if (subCommand.equalsIgnoreCase("levelspells"))
+		{
+			if (!api.hasPermission(sender, "Magic.commands." + command + "." + subCommand)) return true;
+
+			onWandLevelSpells(sender, player, args2);
 			return true;
 		}
 		if (subCommand.equalsIgnoreCase("fill"))
@@ -1091,11 +1099,31 @@ public class WandCommandExecutor extends MagicConfigurableExecutor {
 		sender.sendMessage(message);
 		return true;
 	}
+
+	public boolean onWandLevelSpells(CommandSender sender, Player player, String[] parameters)
+	{
+		Integer maxLevel = null;
+		if (parameters.length > 0) {
+			try {
+				maxLevel = Integer.parseInt(parameters[0]);
+			} catch (Exception ex) {
+				sender.sendMessage("Usage: /wand levelspells <level>");
+				return true;
+			}
+		}
+
+		Wand wand = checkWand(sender, player);
+		if (wand == null) {
+			return true;
+		}
+
+		return onLevelSpells("wand", sender, player, wand, maxLevel);
+	}
 	
 	public boolean onWandAdd(CommandSender sender, Player player, String[] parameters)
 	{
 		if (parameters.length < 1) {
-			sender.sendMessage("Use: /wand add <spell|material> [material:data]");
+			sender.sendMessage("Usage: /wand add <spell|material> [material:data]");
 			return true;
 		}